fix(web-push): keep SW registration out of the initial /chat bundle; raise the chat budget for the feature's i18n copy
The merge-lane bundle-budget gate (skipped on the PR lane) flagged the initial
/chat JS at 218.2 KB against the 217.0 KB gzip budget. `main.tsx` boot-imported
`registerServiceWorker` from `lib/web-push.ts`, which dragged the enrollment
lib's api-client and WebCrypto imports into the initial chunk. Extracted the
dependency-free `lib/register-sw.ts` for boot; the enrollment API stays in
`web-push.ts`, imported only by the already-lazy automations route. That
recovered the eager-code weight (218.2 -> 217.4 KB); the residual is the
feature's new `en.ts` fallback-pack strings, so the /chat budget is raised
217.0 -> 218.0 KB with rationale, matching how prior features (hosted MCP,
Router 8) handled eager localized copy.
