+++ crates/product/ironclaw_webui/frontend/src/main.tsx
@@ -4,7 +4,7 @@ import { createRoot } from "react-dom/client";
 import { App } from "./app/app";
 import { queryClient } from "./lib/query-client";
 import { I18nProvider } from "./lib/i18n";
-import { registerServiceWorker } from "./lib/web-push";
+import { registerServiceWorker } from "./lib/register-sw";
 // Only the English fallback is bundled eagerly; every other locale is
 // lazy-loaded on demand by I18nProvider (see lib/i18n.tsx `loaders`).
 import "./i18n/en";
